type fixed in printf

+++ shmutil.c
@@ -39,7 +39,7 @@ int main(int argc, char **argv) {
 		URN_INFOF("\tshm_segsz:   %zu", ds.shm_segsz);
 		URN_INFOF("\tlast op pid: %d", ds.shm_lpid);
 		URN_INFOF("\tcreator pid: %d", ds.shm_cpid);
-		URN_INFOF("\tnum attach:  %d", ds.shm_nattch);
+		URN_INFOF("\tnum attach:  %lu", (unsigned long)(ds.shm_nattch));
 	} else if (strcasecmp(argv[2], "del") == 0) {
 		rv = shmctl(shmid, IPC_RMID, NULL);
 		if (errno != 0)
